# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/dashboard:
    get:
      operationId: getDashboardMetrics
      tags: [dashboard]
      summary: CRM dashboard metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/learning/dashboard:
    get:
      operationId: learningDashboard
      tags: [learning]
      summary: Learning dashboard metrics
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningDashboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/learning/catalogue:
    get:
      operationId: learningCatalogue
      tags: [learning]
      summary: Training catalogue (M1 content)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Catalogue items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningCatalogueResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/learning/enrolments:
    get:
      operationId: listEnrolments
      tags: [learning]
      summary: List enrolments
      responses:
        "200":
          description: Enrolment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrolmentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createEnrolment
      tags: [learning]
      summary: Create enrolment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrolmentRequest"
      responses:
        "201":
          description: Enrolment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrolment"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/learning/transcript:
    get:
      operationId: learningTranscript
      tags: [learning]
      summary: Learning transcript
      responses:
        "200":
          description: Transcript
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/learning/certificates:
    get:
      operationId: listLearningCertificates
      tags: [learning]
      summary: List course certificates
      responses:
        "200":
          description: Certificates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningCertificateListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/learning/analytics:
    get:
      operationId: learningAnalytics
      tags: [learning]
      summary: Learning analytics
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningAnalyticsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/learning/trainer/dashboard:
    get:
      operationId: trainerDashboard
      tags: [learning]
      summary: Trainer workspace dashboard
      responses:
        "200":
          description: Trainer dashboard
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
...
    Unauthorized:
      description: Authentication is required, or the session has expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: >
        Authenticated, but lacking the required permission or tenant scope.
        Retrying with the same credentials will not succeed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Too many requests. Retry after the window resets.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: >
        An unexpected fault. The message is generic by design; quote
        `error.requestId` when reporting it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: The service is temporarily unable to handle the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DashboardMetrics:
      type: object
      properties:
        newLeads:
          type: number
        qualified:
          type: number
        won:
          type: number
        lost:
          type: number
        openLeads:
          type: number
        averageResponseMinutes:
          type: number
          nullable: true
      required:
        - newLeads
        - qualified
        - won
        - lost
        - openLeads
...
    FieldError:
      type: object
      description: A single field-level validation failure.
      properties:
        path:
          type: string
          description: Dotted path to the offending field, e.g. `contact.email`.
        message:
          type: string
      required:
        - path
        - message
...
    LearningDashboardResponse:
      type: object
      additionalProperties: true
    LearningCatalogueResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - items
    Enrolment:
      type: object
      additionalProperties: true
    EnrolmentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Enrolment"
      required:
        - items
    CreateEnrolmentRequest:
      type: object
      properties:
        courseCategory:
          type: string
        courseSlug:
          type: string
        pathwayId:
          type: string
          nullable: true
        isMandatory:
          type: boolean
        source:
          type: string
      required:
        - courseCategory
        - courseSlug
    LearningCertificateListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - items
    LearningAnalyticsResponse:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      description: >
        Standard error envelope returned by every failing endpoint. The `code`
        enum is kept in lockstep with `ErrorCode` in
        `lib/platform/src/errors`; adding a code requires updating both.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: >
                Stable, machine-readable classification. Clients branch on this
                rather than on the HTTP status or the message text.
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - payload_too_large
                - unprocessable_entity
                - rate_limited
                - internal_error
                - service_unavailable
            message:
              type: string
              description: >
                Human-readable summary. Safe to display for expected errors;
                generic for internal faults, which never disclose their cause.
            details:
              type: object
              description: >
                Structured, client-safe detail. Carries `fieldErrors` for
                validation failures; otherwise error-specific.
              properties:
                fieldErrors:
                  type: array
                  items:
                    $ref: "#/components/schemas/FieldError"
              additionalProperties: true
            requestId:
              type: string
              description: >
                Correlation id, matching the `x-request-id` response header and
                the server logs. Quote it when reporting a fault.
          required:
            - code
            - message
      required:
        - error